# target/classes/openapi.yaml
openapi: 3.0.3
info:
  title: DigiShopKart
  description: Apis of E-commerce application
  contact:
    name: Ajay Kate
    url: http://digishopkart.com
  version: '0.1'
servers:
  - url: http://localhost:8085
#------------------------------PATHS ---------------------------------
paths:
...
  /api/v1/customers:
    post:
      tags:
        - customer
      summary: add a customer
      operationId: addCustomer
      requestBody:
        $ref: '#/components/requestBodies/CustomerBody'
      responses:
        200:
          description: sucess response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        404:
          $ref: '#/components/responses/NotFound'
        400:
          $ref: '#/components/responses/BadRequest'

    get:
      tags:
        - customer
      summary: fetch all customers
      operationId: fetchAllCustomers
      description: all customer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customers/{customerId}:
    get:
      tags:
        - customer
      operationId: fetchCustomer
      summary: fetch a customer
      description: fetching customer By customerId
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - customer
      operationId: deleteCustomer
      summary: delete a customer
      description: api to delete the customer
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: "Data deletion request Successfully Accepted"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
    put:
      tags:
        - customer
      summary: update a customer
      operationId: updateCustomer
      description: update Customer by using Customer_id
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
      requestBody:
        $ref: '#/components/requestBodies/CustomerBody'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'


  #  ****************************** CUSTOMER ADDRESS  *****************************
  /api/v1/customerAddress:
    get:
      tags:
        - CustomerAddress
      summary: fetch all customerAddresses
      operationId: fetchAllCustomerAddresses
      description: API to fetch all customer Addresses
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerAddress'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CustomerAddress
      summary: insert a CustomerAddress
      operationId: insertCustomerAddress
      description: API to insert Customer Address record
      requestBody:
        $ref: '#/components/requestBodies/CustomerAddressBody'
      parameters:
        - in: query
          name: customerId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAddress'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customerAddress/{customerId}:
    get:
      tags:
        - CustomerAddress
      summary: fetch a CustomerAddresses
      operationId: fetchCustomerAddress
      description: API to fetch The Customer Address
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAddress'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customerAddress/{addressId}:
    put:
      tags:
        - CustomerAddress
      summary: update a CustomerAddress
      operationId: updateCustomerAddress
      description: API to update the Customer Address by addressId
      parameters:
        - in: path
          name: addressId
          required: true
          schema:
            type: integer
      requestBody:
        $ref: '#/components/requestBodies/CustomerAddressBody'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAddress'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - CustomerAddress
      summary: delete a CustomerAddress
      operationId: deleteCustomerAddress
      description: API to delete Customer Address by addressId
      parameters:
        - in: path
          name: addressId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAddress'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'


  #  ********************************** PRODUCT  **********************************
...
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
  #___________________REQUEST BODY_____________________________________
  requestBodies:
...
    CustomerBody:
      description: A JSON object Contenting Customer and CustomerAddress information
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Customer'
    CustomerAddressBody:
      description: A JSON object Contenting CustomerAddress information
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CustomerAddress'
...
  schemas:
    Errors:
      type: object
      properties:
        code:
          type: string
        message:
          title: string
      required:
        - code
        - message
    Customer:
      type: object
      required:
        - firstName
        - lastName
        - email
        - mobile
        - customerStatus
        - customerAddress
      properties:
        customerId:
          type: integer
        firstName:
          type: string
          example: Ajay
          pattern: '^[a-zA-Z]+(\s[a-zA-Z]+)?$'
        lastName:
          type: string
          example: Kate
          pattern: '^[a-zA-Z]+(\s[a-zA-Z]+)?$'
        email:
          type: string
          example: ajaykate@gmail.com
          format: email
        mobile:
          type: string
          example: +919876543210
        customerStatus:
          type: string
          enum:
            - active
            - inactive
          default: active
          nullable: false
        customerAddress:
          type: array
          items:
            $ref: '#/components/schemas/CustomerAddress'
    CustomerAddress:
      type: object
      required:
        - customerFullName
        - country
        - state
        - town
        - area
        - houseOrBuilding
        - landmark
        - pinCode
        - mobile
      properties:
        id:
          type: integer
        addressType:
          type: string
          enum:
            - shipping
            - billing
        customerFullName:
          type: string
        country:
          type: string
        state:
          type: string
        town:
          type: string
        area:
          type: string
        houseOrBuilding:
          type: string
        landmark:
          type: string
        pinCode:
          type: integer
        mobile:
          type: string
...